feat: add Spending History Section components and hooks
- Implement SpendingHistorySummaryCards for displaying summary of spending by state and totals.
- Create SpendingHistoryTable for detailed transaction view with expandable categories and subcategories.
- Introduce TransactionDetailTable for displaying individual transaction details.
- Add utility functions and constants for formatting and managing spending data.
- Implement hooks for managing state and data in the Spending History Section.
